feat: Add build/ directory to .gitignore
Added the `build/` directory to the `.gitignore` file to exclude it from version control.

Update meta description in index.html

Updated the meta description in the `index.html` file to reflect the purpose of the website as "A easy, scalable and scalable message hub for event on TON".

Update page title in index.html

Changed the page title in the `index.html` file to "Eventon" for better branding.